use github-url-to-object
This patches removes `github-url` in favor of `github-url-to-object`, the same
module that the new npmjs website uses to parse repository URLs.

Repository URLs not ending with ".git", like `review`'s:

    $ npm view review repository.url
    git://github.com/juliangruber/review

cannot be parsed by `github-url` (although there is a pull request open for
this), so ironically enough the example <http://ghub.io/review> goes to the
npmjs page.

In addition, packages using the new GitHub shorthand notation (`user/repo`)
expand at publish-time to something that `github-url` cannot parse at all:

    $ npm view binomial-cdf repository.url
    git+https://github.com/kenany/binomial-cdf.git

But `github-url-to-object` parses all of these.